Importing more complete word lists
The previous ones were somwhat limited, which resulted in some messages
getting a neutral rating. These word lists are from a sentiment analysis
project on Twitter data. The advantage of this is that it also contains
informal words (which developers tend to use in commit messages) as well
as some commonly misspelled words that also appear in commit messages.