Make analyze.py only accept preprocess.py's output
Run it like this:

$ python analyze.py < commit_comments.json

It will consider each message (JSON object) on its own and print the
calculated sentiment score.

There are some commit comments that are not helpful (non-English, et
cetera), therefore for the classifier we might end up having to compose
a custom dataset of helpful commit comments with manually set scores.
We'll see, but for now this is working just fine. Yay ^_^