Added message display with word scores coloring
In analyze.py, display the messages on one line by stripping newlines.
Also show the words that affected the score with a respective color in
front of it.

Best way to run this is:
$ python analyze.py < commit_comments.json | less -SR
This makes less not wrap lines and show the colors correctly.

Later on with real classifiers we can disable/remove this display since
it will probably not be needed then.