Call predict for every message to avoid memory hog
Calling the prediction function first means we have to go through the
whole data set before we can display the predictions, which means
storing a bunch of explicit lists. This way we can avoid the memory
usage.

The prediction function shouldn't be too slow to call every time (still
have to try out in practice...). It also looks cleaner.

Also made the read_json iterator yield only the message and relevant
group, which reduced memory a little bit.

This basically makes the Analyzer more aware of the group parameter and
cleans up the code some more as well.