Hybrid iterator mapping method that extracts lists
Implement a way to still use iterators where possible, falling back to
lists where this is sane to do.

In the optimal case, we pass an iterator to the predictor because we
don't need to display anything, and we skip creating a list of groups
because we use the scores themselves as groups.

However, we must also be able to work with other situations, like when
we want a display and grouping. Thus we make lists for them in those
situations.

It seems from some benchmarking that most of the memory usage comes
from the batch prediction, so that is not too problematic to add lists
in those cases. The optimal case for the hybrid method does save memory
but not as much as hoped for, but at least it does not degrade time.

Prediction times without output (classify.py, score group, >/dev/null)
Averaged over two runs on node02

Calling predict every time (bad): Time: 8:26 m (36.6 MB)
Throwing away all other data (not useful): 16.18 s (205 MB)
The naive method (everything is a list): 16.23 s (207 MB)
Iterator map method (optimal hybrid method): 16.50 s (128 MB)

Comparison with early return before predict: 35.6 MB
Comparison with full run of naive analyze.py: 1:00 m (7.4 MB)