Tune labeler to not store so much data
We can get all the data we want from the original dataset by getting it
from the same line number. For the labeled dataset we only need to store
the label and perhaps the ID for the sake of completeness.

Also relabeled the data from yesterday due to these changes and added
some more labels.