Increase the number of estimator trees in classify
This makes full use of the available cores, otherwise each thread or
process will do not as much as it has to.

For some reason, scikit-learn RandomForestRegressor uses the threading
backend of the Parallel module, claiming that it is as fast as
multiprocessing because the tree operations are written in a C library
which overcomes the Global Interpreter Lock that python threads have to
suffer. Still, in one process it is possible to use all cores by
increasing the number of trees, which get divided more usefully.