Added preprocess script for model data
This makes use of the sqlparse library to extract the project and
language from a SQL dump of GHTorrent metadata, or a part of it
containing only the projects table values.

Right now this script is very slow (possibly because of the sqlparse
module), but it seems to be acceptable in terms of memory. We want to
merge this with the preprocess script to conditionally fetch the project
language.

Note that this requires the git master of sqlparse, which contains a fix
to work with escaped backslashes, which throws of the pip version. The
readme notes how to get it.