Removing code for preprocessing MySQL dumps
The MySQL dumps are too large to handle in terms of both time and
memory. I just found out that the language field is also present in the
'repos' BSON dump that GHTorrent provides, so let's use that instead as
it is much easier to handle and we can reuse our current BSON conversion
and filtering logic.

(Sorry for the work that you put into the previous solution, Leon!)