Made the preprocessor more generic
Split the current preprocessor into a base class and a subclass. The
base class contains methods that all preprocessors have in common, such
as those for downloading and untarring the dataset. The subclass
Commit_Comments_Preprocessor inherits the methods from the base class
and adds its own logic specifically for preprocessing the commit
comments dataset.

This split has been made to make the preprocessor code more extendable,
which we will need for the next step: preprocessing the 'repos' BSON
dump.