Implemented multiprocessing for shelves
Run with the following commands:

- mpiexec -n 16 python preprocess.py repos language (make sure that you
  have more than 150 GB of disk space!)
- python preprocess.py commit_comments language

This will be much faster than the previous method we used. Open
problems:

- Scrollbar is broken when using multiprocessing (processes overwrite
  each other's progress bars).
- I only tested it with "-n 2" or slightly more, but since I only have
  80 GB of free space, I could not test the entire "-n 16". We will have
  to do that either on the DAS or somehow group processes.

Other than that, this appears to work like a charm! We have yet to time
this precisely, but initial tests show that this is in fact *a lot*
faster than the previous method ^_^