Automatically balance over processes
Instead of requiring that the same number of processes is run as there
are datasets to download, we simply have a number of processes, and each
time one of them sends a message saying that they are ready, the master
sends another job to it. Thus there are no longer processes waiting to
get a job.

Furthermore, one automatically gets some load balancing: a process that
has done a small data set can quickly get a new one while another
process is still running on a bigger one, and all processes might still
finish at a similar time.

The master was also being idle in between job scheduling, using 100% CPU
just to poll for new messages. Prevent this by using nonblocking send
and receive requests, which allows to use a wait loop that can sleep.
This might make responses by the master a little slower, but that is
negligible compared to the preprocessing, and it reduces CPU a lot.

The MPI run of python only stops when all processes are done, thus one
needs to ensure all of them quit out of their loops. This is done by
going through the master loop one more time and notifying the processes
that they do not get a new job.

Also added some prints to make the communication between master and
processes clear, and added a usage message. One can now run the script
as follows:

mpirun -n <num_processes> python preprocess.py repos language

And the master scheduling automatically uses the correct number of
processes, which reduces the number of arguments. Also, the script can
now be run sequentially again:

python preprocess.py repos language