The Return of unrecognized.py
Added a way to compare the analyzer with the classifier by comparing
whether their scores are somewhat alike. If the analyzer claims
a message is negative while the classifier thinks it is positive, then
probably one of the models was not trained well, and the message might
be important to make it better.

The only thing is that this seems to yield very similar messages for the
same score difference, such as bot messages and messages like "fix bug"
which don't have much additional use. We might want to compare
differently or group such messages, in order to provide the most useful
messages for the labeler, for example.