preprocess: Fixed theoretic skipping of jobs
The master could potentially skip jobs if it received a ready message
different from a byte '\1'. While the current code never causes this, it
is better to check correctly before continuing.

Furthermore, there was a stray `tag = tag + 1` which might have made it
skip jobs. This is a slightly new bug since the outer loop was
previously a while loop, and this statement was necessary to continue in
the loop. It should have been removed in commit 189e9d5, but it was not.
Note that Python ignores the change to the tag variable and continues
the loop as usual, since it is extracting from an iterator which
preserves the iteration we are in.