Classify: Added cross-validation option
To perform cross-validation, run with python classify.py score 100 5

The last parameter is the number of folds (the one before that is the
number of trees in RandomForest), which replaces the model file when it
is a number instead of a file name. It is not recommended to run
cross-validation on an already trained model (since it then knows
everything about the labeled dataset), so that is prevented in this way.

Skip the prediction phase and other parts of the classifier when in
cross-validation mode for now.