Classify: Tokenize using analyzer's message split
Refactor Analyze.split to Utilities.split, now with a class constant for
the large regex for splitting boundaries and preserving emoticons. Thus
we can now use it in the classifier, which can be easily done by giving
it to the tokenizer parameter of the TfidfVectorizer. The only problem
is that it cannot be pickled, so it needs to be removed when creating
the pickle and readded when loading it.

From the results of cross-validations, it appears that this makes the
classifier faster for some reason (perhaps because the tree is smaller
according to tree.py). Furthermore, the average score is about 0.1
higher (from about 0.32 to 0.42), which is a very nice gain for a small
change.