Implemented experiment runner
Run with `python experiment.py [folds]`. This program allows us to
quickly test different classifiers or regressors that scikit-learn
provides.

The classifiers that are in the list right now are the ones that we can
actually run. There are more on the scikit-learn website, but those
complain about our sparse matrices.