classify.py: Find all commit comments dumps
If no file is given on standard input, then search the current directory
and subdirectories of process IDs for commit_comments dump files. We
have to use a specific pattern in order to exclude the labeled data set.
The files are then passed one by one to the predict function which can
now handle any (file) stream.