Added language plots for all commit comments dumps
These plots show the "relative scores" of the top 20 languages sorted on
negative and positive scores, respectively, based on the classifications
of the RandomForestRegressor on all commit comments from the history of
dumps of GHTorrent.

The relative score, as mentioned before, takes into account the size of
each language group as well as the weighted scores of each example. The
latter fact is the reason we use a regressor for this; a classifier
should also work but might be harder to compare results for due to the
discrete -1/0/1 labels.

The file names are not that well chosen. In case we want to do more runs
with other classifiers or regressor versions, then we can always rename
these.

Interesting results: Rust is very positive, although this is probably
related to bot outputs (but that means their tests often pass!), as is
Erlang and Vim. Scala and Matlab languages have very negative
communities. At least the latter is probably in line of expectations.