Presentation: most of the slides
I think we only need to add more about experiments and a conclusion
really. We only have 20 minutes and this should be more than enough to
fill those 20 minutes.

For the implementation part I have made a timeline-like view because
the events are chronological. The timeline has been composed by manually
going through the Git commit history and listing the most important
points. I think this way we provide a nice chronological overview of our
implementation. We will have to spend quite some time on the
implementation and the experiments as those parts are relatively new for
the group and are the most important.