Replace group size reweighting with a filter
Simplify the meaning of the relative score by removing the group size
ratio from the equation. Instead filter the pandas data frame by only
using groups that have enough size based on the maximum score count.

Also fixed a bug which made the input not always detect that we want to
make a plot sorted on negatives.