Updated numbers on Dataset slide
There was an inconsistency between the given number for the 2015-01-29
commit comments dump: there are 182.282 comments in total in that dump,
but we only take 165.427 of them by filtering non-Latin comments.

Note that 2015-01-29 is no longer the "last" dump, 2015-03-29 is
currently the newest.

Also added statistics gathered for all the 17 dumps (using some MPI and
bash magic) as well as for the repos dumps. Made the example field a bit
smaller to make everything fit.