Consider presentation and report as documentation
This should make the language statistics on GitHub not think that half
of the repository is LaTeX, thus making the detection more exact.
I verified this locally with their github-linguist ruby gem, which was
problematic at first but it seems to only update after an actual commit
(so this is an amendment).