PassiveAggressive language plots in presentation
These plots are from a run on all commit comments with languages using
the PassiveAggressive regressor (in order to receive scaled ratings)
with C=0.4. There are some more small languages in these plots compared
to the RandomForest one, but it's still interpretable.

Kept the RandomForest ones for reference, with some better names.