README: Added overview of commands
We have quite a few scripts, and not all of them were explained in the
README. It is also nice to have an overview of which script does what,
and whether they can be run with MapReduce or MPI.

This is also extremely useful for the implementation section of the
report, where we should describe step by step what we implemented and
why. This goes in concert with the timeline from the presentation, with
some more details about the scripts.