Added timing experiments
Initial results of timing experiments on the DAS-3 fileserver for runs
of our scripts without any MapReduce or MPI, and with lots of disk I/O
and networking lag.

Also added a comparison of preprocess.py without anything and
preprocess.py with the dumps already downloaded; note how the user and
system time has almost no difference. Also for the repos task, the disk
I/O only counts in the real time and not for the user and sys time,
since the run time of 21 hours is 2.5 times as much as the user time.

This shows why we should use the user+sys time everywhere. Still have to
figure out how to get timing working on MapReduce (perhaps we can sneak
memtime.sh in there); on MPI we could run memtime.sh for every job and
collect all the results.

Made memtime.sh separate its own output and the script output into
STDERR and STDOUT, so that the commands can be combined and output
redirected separately. For example, to time the emulation of MapReduce:

./memtime.sh python classify.py score < ...json | ./memtime.sh sort \
 | ./memtime python reducer.py score > /dev/null