Report: Tables for timing and memory experiments
These contain the results of the timing experiments, formatted in
a nicer way and with some more calculations than in timing.txt.
Hopefully the structure of component-based and framework-based
comparisons are somewhat readable.

The numbers are somewhat difficult to compare due to different targets
in the experiments (for example, the MPI run performs on all 17 dumps
whereas others only use one dump without languages, etc.), and because
of the differences in timing (user/system time is summed for the
distributed platforms, which makes it seem weird compared to the faster
real time).

Added some TODO-like notes as a basis for explanations in the report.